Add `demo.html` file to the root of our project.
Let's start building out the conventional folder structure Polymer projects usually follow:

    minesweeper/
      -bower_components/
      -bower.json
      -demo.html

We've already taken care of the `bower` parts of the directory structure, so the main thing that remains is the canonical `demo.html` file, which presents a demonstration of the web components defined in this package. We don't have any yet but we'll create the file as a placeholder for now.

### Viewing our progress using polyserve

To prepare you for building our first component, it's important that you understand how to view your components in a browser during development. This process used to be harder because of Polymer's unique folder structure but is not made much simpler using a tool built by a Polymer team member: Polyserve. Polyserve works by serving your package and its Bower dependencies at artificial HTTP paths that make it easier for components to find and include each other in the browser.

> For a thorough explanation of the folder structure complications I mentioned before, check out [Polycast #14](https://www.youtube.com/watch?v=p7Q1mQtFGM8)

Installing [Polyserve](https://github.com/PolymerLabs/polyserve) is pretty easy:

    npm install polyserve -g

> If `npm` is missing on your system you'll need to install [Node.js](https://nodejs.org/)

With Polyserve installed we can load up our `demo.html` page in a browser:

    ➜  minesweeper git:(1-starting-polymer-project) ✗ polyserve
    Starting Polyserve on port 8080
    Files in this directory are available at localhost:8080/components/minesweeper/...

Now to see our `demo.html` file in a browser we visit: http://localhost:8080/components/minesweeper/demo.html

![demo.html running in Chrome](1-starting-polymer-project/1.png)