## Test-driven development
We will be building our first component using a test-driven development approach. Polymer comes with a great testing framework that we previously installed: `web-component-tester`, often abbreviated as WCT. In the coming few sections we'll learn how to setup WCT, write and run our tests, and learn a workflow for building Polymer components this way.

The convention for testing Polymer components is to create a subfolder in your package named `test`, which contains an HTML file for each component in your package. For example:

    test/
      -mine-cell.html

For now we'll just create the file as HTML boilerplate for the moment and then fill it out step by step with explanation.